Added hash_salt, and changed admin username
The addition of the `hash_salt` setting to the `docker_entrypoint.sh`
script is not optimal as this is going to be retained in the DB.
However, since the data in the `seed.sql` file is dependent on the
correct hash in settings, we must do this if we want to use the seed
data to initialize our project and avoid a new install.

In a production environment this salt hash will be changed and as long
as the seed data doesn't contain and private data from production, the
risk is minimal.