Removed some Drupal core files
The settings in .gitattributes look useful but I haven't had time to see
if this will resolve the line ending issues we describe in the
`README.md` for this project. I will perhaps take a look at that in the
future and see if we can resolve things from the linux scripts as well!